add UT and fix bug

+++ pyzoo/zoo/automl/model/base_pytorch_model.py
@@ -54,11 +54,11 @@ class PytorchBaseModel(BaseModel):
 
     @staticmethod
     def covert_input(x, y, validation_data):
-        x = PytorchBaseModel.to_torch(x)
-        y = PytorchBaseModel.to_torch(y)
+        x = PytorchBaseModel.to_torch(x).float()
+        y = PytorchBaseModel.to_torch(y).float()
         if validation_data is not None:
-            validation_data = (PytorchBaseModel.to_torch(validation_data[0]),
-                               PytorchBaseModel.to_torch(validation_data[1]))
+            validation_data = (PytorchBaseModel.to_torch(validation_data[0]).float(),
+                               PytorchBaseModel.to_torch(validation_data[1]).float())
         return x, y, validation_data
 
     def _train_epoch(self, x, y):
